feat: move tracer to its own package and create metrics
