+++ components/ledger/pkg/opentelemetry/metrics/metrics.go
@@ -1,0 +1,1 @@
+package metrics
